+++ contracts/test/PriceMathEchidnaTest.sol
@@ -13,13 +13,13 @@ contract PriceMathEchidnaTest {
     using SafeMath for uint256;
 
     function getAmountOutInvariants(
-        uint112 reserveIn,
-        uint112 reserveOut,
-        uint112 amountIn
+        uint256 reserveIn,
+        uint256 reserveOut,
+        uint256 amountIn
     ) external pure {
         require(reserveIn > 0 && reserveOut > 0);
 
-        uint112 amountOut = PriceMath.getAmountOut(reserveIn, reserveOut, amountIn);
+        uint256 amountOut = PriceMath.getAmountOut(reserveIn, reserveOut, amountIn);
         assert(amountOut < reserveOut);
 
         uint256 k = uint256(reserveIn).mul(reserveOut);
@@ -29,16 +29,16 @@ contract PriceMathEchidnaTest {
         assert(kAfter >= k);
     }
 
-    function roundingCanBeGreaterThan1(uint224 price, uint256 liquidity) external pure {
+    function roundingCanBeGreaterThan1(uint224 price, uint112 liquidity) external pure {
         require(price >= TickMath.getRatioAtTick(TickMath.MIN_TICK)._x);
         require(price <= TickMath.getRatioAtTick(TickMath.MAX_TICK)._x);
 
-        (uint112 amount0Up, uint112 amount1Up) = PriceMath.getVirtualReservesAtPrice(
+        (uint256 amount0Up, uint256 amount1Up) = PriceMath.getVirtualReservesAtPrice(
             FixedPoint.uq112x112(price),
             liquidity,
             true
         );
-        (uint112 amount0Down, uint112 amount1Down) = PriceMath.getVirtualReservesAtPrice(
+        (uint256 amount0Down, uint256 amount1Down) = PriceMath.getVirtualReservesAtPrice(
             FixedPoint.uq112x112(price),
             liquidity,
             false
@@ -60,14 +60,14 @@ contract PriceMathEchidnaTest {
         require(lpFee > 0 && lpFee < PriceMath.LP_FEE_BASE);
 
         FixedPoint.uq112x112 memory price = FixedPoint.uq112x112(priceRaw);
-        (uint112 reserve0, uint112 reserve1) = PriceMath.getVirtualReservesAtPrice(price, liquidity, false);
+        (uint256 reserve0, uint256 reserve1) = PriceMath.getVirtualReservesAtPrice(price, liquidity, false);
 
         require(reserve0 > 0 && reserve1 > 0);
 
         FixedPoint.uq112x112 memory priceTarget = TickMath.getRatioAtTick(tickTarget);
         bool zeroForOne = price._x >= priceTarget._x;
 
-        (uint112 amountIn, uint112 amountOutMax) = PriceMath.getInputToRatio(
+        (uint256 amountIn, uint256 amountOutMax) = PriceMath.getInputToRatio(
             reserve0,
             reserve1,
             liquidity,
@@ -87,12 +87,12 @@ contract PriceMathEchidnaTest {
             uint112 amountInLessFee = uint112(
                 (uint256(amountIn) * (PriceMath.LP_FEE_BASE - lpFee)) / PriceMath.LP_FEE_BASE
             );
-            uint112 amountOut = zeroForOne
+            uint256 amountOut = zeroForOne
                 ? PriceMath.getAmountOut(reserve0, reserve1, amountInLessFee)
                 : PriceMath.getAmountOut(reserve1, reserve0, amountInLessFee);
 
             // downward-adjust amount out if necessary
-            amountOut = uint112(Math.min(amountOut, amountOutMax));
+            amountOut = Math.min(amountOut, amountOutMax);
 
             // (uint112 reserve0Next, uint112 reserve1Next) = zeroForOne
             //     ? (reserve0 + amountInLessFee, reserve1 - amountOut)
